pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-document-processor"
version = "0.1.0"
description = "Microserviço para processamento de documentos financeiros com IA"
readme = "README.md"
authors = [
    {name = "Seu Nome", email = "seu.email@exemplo.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "aiokafka>=0.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
    "loguru>=0.7.0",
    "httpx>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.5",
    "asyncpg>=0.27.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.5.0",
    "pypdf>=3.15.1",
    "pytesseract>=0.3.10",
    "tenacity>=8.2.0",
    "prometheus-client>=0.17.0",
    "alembic>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.260",
    "greenlet>=3.2.2"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]